uefi

Module boot

UEFI boot services.

These functions will panic if called after exiting boot services.

Accessing protocols

Protocols can be opened using several functions in this module. Most commonly, open_protocol_exclusive should be used. This ensures that nothing else can use the protocol until it is closed, and returns a ScopedProtocol that takes care of closing the protocol when it is dropped.

Other methods for opening protocols:

open_protocol

get_image_file_system

For protocol definitions, see the proto module.

Structs

EventType

Flags describing the type of an UEFI event and its attributes.

HandleBuffer

A buffer returned by locate_handle_buffer that contains an array of Handles that support the requested protocol.

MemoryAttribute

Flags describing the capabilities of a memory range.

MemoryDescriptor

A structure describing a region of memory. This type corresponds to version of this struct in the UEFI spec and is always bound to a corresponding UEFI memory map.

MemoryType

The type of a memory range.

OpenProtocolParams

Parameters passed to open_protocol.

ProtocolSearchKey

Opaque pointer returned by register_protocol_notify to be used with locate_handle via SearchType::ByRegisterNotify.

ProtocolsPerHandle

Protocol interface Guids that are installed on a Handle as returned by protocols_per_handle.

ScopedProtocol

An open protocol interface. Automatically closes the protocol interface on drop.

Tpl

Task priority level.

TplGuard

RAII guard for task priority level changes.

Enums

AllocateType

Type of allocation to perform.

LoadImageSource

Used as a parameter of load_image to provide the image source.

OpenProtocolAttributes

Attributes for open_protocol.

SearchType

The type of handle search to perform.

TimerTrigger

Trigger type for events of type TIMER.

Constants

PAGE_SIZE

Size in bytes of a UEFI page.

Functions

allocate_pages

Allocates memory pages from the system.

allocate_pool

Allocates from a memory pool. The pointer will be 8-byte aligned.

check_event

Checks to see if an event is signaled, without blocking execution to wait for it.

close_event

Removes event from any event group to which it belongs and closes it.

connect_controller

Connect one or more drivers to a controller.

create_event⚠

Creates an event.

create_event_ex⚠

Creates an event in an event group.

disconnect_controller

Disconnect one or more drivers from a controller.

exit⚠

Exits the UEFI application and returns control to the UEFI component that started the UEFI application.

exit_boot_services⚠

Exit UEFI boot services.

free_pages⚠

Frees memory pages allocated by allocate_pages.

free_pool⚠

Frees memory allocated by allocate_pool.

get_handle_for_protocol

Find an arbitrary handle that supports a particular Protocol. Returns NOT_FOUND if no handles support the protocol.

get_image_file_system

Retrieves a SimpleFileSystem protocol associated with the device the given image was loaded from.

image_handle

Get the Handle of the currently-executing image.

install_configuration_table⚠

Adds, updates, or removes a configuration table entry from the EFI System Table.

install_protocol_interface⚠

Installs a protocol interface on a device handle.

load_image

Loads a UEFI image into memory and return a Handle to the image.

locate_device_path

Locates the handle of a device on the device path that supports the specified protocol.

locate_handle

Enumerates all handles installed on the system which match a certain query.

locate_handle_buffer

Returns an array of handles that support the requested protocol in a pool-allocated buffer.

memory_map

Stores the current UEFI memory map in an UEFI-heap allocated buffer and returns a MemoryMapOwned.

open_protocol⚠

Opens a protocol interface for a handle.

open_protocol_exclusive

Opens a protocol interface for a handle in exclusive mode.

protocols_per_handle

Get the list of protocol interface Guids that are installed on a Handle.

raise_tpl⚠

Raises a task’s priority level and returns a TplGuard.

register_protocol_notify

Registers event to be signaled whenever a protocol interface is registered for protocol by install_protocol_interface or reinstall_protocol_interface.

reinstall_protocol_interface⚠

Reinstalls a protocol interface on a device handle. old_interface is replaced with new_interface. These interfaces may be the same, in which case the registered protocol notifications occur for the handle without replacing the interface.

set_image_handle⚠

Update the global image Handle.

set_timer

Sets the trigger for an event of type TIMER.

set_watchdog_timer

Sets the watchdog timer.

stall

Stalls execution for the given number of microseconds.

start_image

Transfers control to a loaded image’s entry point.

test_protocol

Tests whether a handle supports a protocol.

uninstall_protocol_interface⚠

Removes a protocol interface from a device handle.

unload_image

Unloads a UEFI image.

wait_for_event

Stops execution until an event is signaled.

Type Aliases

EventNotifyFn

Event notification callback type.