uefi::boot

Function create_event

pub unsafe fn create_event(
    event_ty: EventType,
    notify_tpl: Tpl,
    notify_fn: Option<EventNotifyFn>,
    notify_ctx: Option<NonNull<c_void>>,
) -> Result<Event>

Creates an event.

This function creates a new event of the specified type and returns it.

Events are created in a “waiting” state, and may switch to a “signaled” state. If the event type has flag NotifySignal set, this will result in a callback for the event being immediately enqueued at the notify_tpl priority level. If the event type has flag NotifyWait, the notification will be delivered next time wait_for_event or check_event is called. In both cases, a notify_fn callback must be specified.

Safety

This function is unsafe because callbacks must handle exit from boot services correctly.

Errors

Status::INVALID_PARAMETER: an invalid combination of parameters was provided.

Status::OUT_OF_RESOURCES: the event could not be allocated.