uefi::boot

Function exit_boot_services

pub unsafe fn exit_boot_services(memory_type: MemoryType) -> MemoryMapOwned

Exit UEFI boot services.

After this function completes, UEFI hands over control of the hardware to the executing OS loader, which implies that the UEFI boot services are shut down and cannot be used anymore. Only UEFI configuration tables and run-time services can be used.

The memory map at the time of exiting boot services returned. The map is backed by a pool allocation of the given memory_type. Since the boot services function to free that memory is no longer available after calling exit_boot_services, the allocation will not be freed on drop.

Note that once the boot services are exited, associated loggers and allocators can’t use the boot services anymore. For the corresponding abstractions provided by this crate (see the helpers module), invoking this function will automatically disable them. If the global_allocator feature is enabled, attempting to use the allocator after exiting boot services will panic.

Safety

The caller is responsible for ensuring that no references to boot-services data remain. A non-exhaustive list of resources to check:

All protocols will be invalid after exiting boot services. This includes the Output protocols attached to stdout/stderr. The caller must ensure that no protocol references remain.

The pool allocator is not usable after exiting boot services. Types such as PoolString which call free_pool on drop must be cleaned up before calling exit_boot_services, or leaked to avoid drop ever being called.

All data in the memory map marked as MemoryType::BOOT_SERVICES_CODE and MemoryType::BOOT_SERVICES_DATA will become free memory.

Errors

This function will fail if it is unable to allocate memory for the memory map, if it fails to retrieve the memory map, or if exiting boot services fails (with up to one retry).

All errors are treated as unrecoverable because the system is now in an undefined state. Rather than returning control to the caller, the system will be reset.