uefi::boot

Function load_image

pub fn load_image(
    parent_image_handle: Handle,
    source: LoadImageSource<'_>,
) -> Result<Handle>

Loads a UEFI image into memory and return a Handle to the image.

There are two ways to load the image: by copying raw image data from a source buffer, or by loading the image via the SimpleFileSystem protocol. See LoadImageSource for more details of the source parameter.

The parent_image_handle is used to initialize the parent_handle field of the LoadedImage protocol for the image.

If the image is successfully loaded, a Handle supporting the LoadedImage and LoadedImageDevicePath protocols is returned. The image can be started with start_image and unloaded with unload_image.

Errors

Status::INVALID_PARAMETER: source contains an invalid value.

Status::UNSUPPORTED: the image type is not supported.

Status::OUT_OF_RESOURCES: insufficient resources to load the image.

Status::LOAD_ERROR: the image is invalid.

Status::DEVICE_ERROR: failed to load image due to a read error.

Status::ACCESS_DENIED: failed to load image due to a security policy.

Status::SECURITY_VIOLATION: a security policy specifies that the image should not be started.