uefi::boot

Function open_protocol

pub unsafe fn open_protocol<P: ProtocolPointer + ?Sized>(
    params: OpenProtocolParams,
    attributes: OpenProtocolAttributes,
) -> Result<ScopedProtocol<P>>

Opens a protocol interface for a handle.

See also open_protocol_exclusive, which provides a safe subset of this functionality.

This function attempts to get the protocol implementation of a handle, based on the protocol GUID.

See OpenProtocolParams and OpenProtocolAttributes for details of the input parameters.

If successful, a ScopedProtocol is returned that will automatically close the protocol interface when dropped.

Safety

This function is unsafe because it can be used to open a protocol in ways that don’t get tracked by the UEFI implementation. This could allow the protocol to be removed from a handle, or for the handle to be deleted entirely, while a reference to the protocol is still active. The caller is responsible for ensuring that the handle and protocol remain valid until the ScopedProtocol is dropped.

Errors

Status::INVALID_PARAMETER: an invalid combination of params and attributes was provided.

Status::UNSUPPORTED: the handle does not support the protocol.

Status::ACCESS_DENIED or Status::ALREADY_STARTED: the protocol is already open in a way that is incompatible with the new request.