uefi::boot

Function register_protocol_notify

pub fn register_protocol_notify(
    protocol: &'static Guid,
    event: &Event,
) -> Result<SearchType<'static>>

Registers event to be signaled whenever a protocol interface is registered for protocol by install_protocol_interface or reinstall_protocol_interface.

If successful, a SearchType::ByRegisterNotify is returned. This can be used with locate_handle or locate_handle_buffer to identify the newly (re)installed handles that support protocol.

Events can be unregistered from protocol interface notification by calling close_event.

Errors

Status::OUT_OF_RESOURCES: the event could not be allocated.