uefi::boot

Function start_image

pub fn start_image(image_handle: Handle) -> Result

Transfers control to a loaded image’s entry point.

Errors

Status::INVALID_PARAMETER: image_handle is not valid, or the image has already been initialized with start_image.

Status::SECURITY_VIOLATION: a security policy specifies that the image should not be started.