uefi::boot

Function wait_for_event

pub fn wait_for_event(events: &mut [Event]) -> Result<usize, Option<usize>>

Stops execution until an event is signaled.

This function must be called at priority level Tpl::APPLICATION.

The input Event slice is repeatedly iterated from first to last until an event is signaled or an error is detected. The following checks are performed on each event:

If an event is of type NOTIFY_SIGNAL, then a Status::INVALID_PARAMETER error is returned with the index of the event that caused the failure.

If an event is in the signaled state, the signaled state is cleared and the index of the event that was signaled is returned.

If an event is not in the signaled state but does have a notification function, the notification function is queued at the event’s notification task priority level. If the execution of the event’s notification function causes the event to be signaled, then the signaled state is cleared and the index of the event that was signaled is returned.

To wait for a specified time, a timer event must be included in events.

To check if an event is signaled without waiting, an already signaled event can be used as the last event in the slice being checked, or the check_event interface may be used.

Errors

Status::INVALID_PARAMETER: events is empty, or one of the events of of type NOTIFY_SIGNAL.

Status::UNSUPPORTED: the current TPL is not Tpl::APPLICATION.