uefi

Module data_types

Data type definitions

This module defines the basic data types that are used throughout uefi-rs

Re-exports

pub use chars::Char16;

pub use chars::Char8;

Modules

chars

UEFI character handling

Structs

CStr8

A null-terminated Latin-1 string.

CStr16

An UCS-2 null-terminated string slice.

Event

Handle to an event structure, guaranteed to be non-null.

Guid

Globally-unique identifier.

Handle

Opaque handle to an UEFI entity (protocol, image…), guaranteed to be non-null.

UnalignedSlice

Slice backed by a potentially-unaligned pointer.

Enums

FromSliceWithNulError

Error converting from a slice (which cannot contain interior nuls) to a string type.

FromStrWithBufError

Error returned by CStr16::from_str_with_buf.

UnalignedCStr16Error

Error returned by CStr16::from_unaligned_slice.

Traits

Align

Trait for querying the alignment of a struct.

EqStrUntilNul

The EqStrUntilNul trait helps to compare Rust strings against UEFI string types (UCS-2 strings). The given generic implementation of this trait enables us that we only have to implement one direction (left.eq_str_until_nul(&right)) for each UEFI string type and we get the other direction (right.eq_str_until_nul(&left)) for free. Hence, the relation is reflexive.

Identify

Several entities in the UEFI specification can be referred to by their GUID, this trait is a building block to interface them in uefi-rs.

Type Aliases

PhysicalAddress

Physical memory address. This is always a 64-bit value, regardless of target platform.

VirtualAddress

Virtual memory address. This is always a 64-bit value, regardless of target platform.