uefi::data_types

Struct UnalignedSlice

pub struct UnalignedSlice<'a, T: Copy> { /* private fields */ }

Slice backed by a potentially-unaligned pointer.

This wrapper can be used to safely expose slices that are inside a repr(packed) struct. The element type must be Copy.

Implementations

impl UnalignedSlice<'_, u16>

pub fn to_cstr16<'buf>( &self, buf: &'buf mut [MaybeUninit<u16>], ) -> Result<&'buf CStr16, UnalignedCStr16Error>

Create a CStr16 from an UnalignedSlice using an aligned buffer for storage. The lifetime of the output is tied to buf, not self.

impl<'a, T: Copy> UnalignedSlice<'a, T>

pub const unsafe fn new(data: *const T, len: usize) -> Self

Create an UnalignedSlice from a raw pointer. The pointer must not be dangling but can be unaligned. The len parameter is the number of elements in the slice (not the number of bytes).

Safety

The data pointer must point to a packed array of at least len elements of type T. The pointer must remain valid for as long as the 'a lifetime.

pub const fn is_empty(&self) -> bool

Returns true if the slice has a length of 0.

pub const fn as_ptr(&self) -> *const T

Get the underlying pointer, which may be unaligned.

pub const fn len(&self) -> usize

Returns the number of elements in the slice.

pub const fn get(&self, index: usize) -> Option<T>

Returns the element at index, or None if the index is out of bounds.

pub const fn iter(&'a self) -> UnalignedSliceIter<'a, T>

Returns an iterator over the slice.

The iterator yields all items from start to end.

pub fn copy_to(&self, dest: &mut [T])

Copy the data to an aligned buffer.

The length of dest must be the same as self.

Panics

This function will panic if the two slices have different lengths.

pub fn copy_to_maybe_uninit(&self, dest: &mut [MaybeUninit<T>])

Copy the data to an aligned MaybeUninit buffer.

The length of dest must be the same as self.

This function fully initializes the dest slice.

Panics

This function will panic if the two slices have different lengths.