uefi::data_types

Trait EqStrUntilNul

pub trait EqStrUntilNul<StrType: ?Sized> {
    // Required method
    fn eq_str_until_nul(&self, other: &StrType) -> bool;
}

The EqStrUntilNul trait helps to compare Rust strings against UEFI string types (UCS-2 strings). The given generic implementation of this trait enables us that we only have to implement one direction (left.eq_str_until_nul(&right)) for each UEFI string type and we get the other direction (right.eq_str_until_nul(&left)) for free. Hence, the relation is reflexive.

Required Methods

fn eq_str_until_nul(&self, other: &StrType) -> bool

Checks if the provided Rust string StrType is equal to Self until the first null character is found. An exception is the terminating null character of Self which is ignored.

As soon as the first null character in either &self or other is found, this method returns. Note that Rust strings are allowed to contain null bytes that do not terminate the string. Although this is rather unusual, you can compare "foo\0bar" with an instance of Self. In that case, only foo" is compared against Self (if Self is long enough).

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".