uefi::data_types

Trait Identify

pub unsafe trait Identify {
    const GUID: Guid;
}

Several entities in the UEFI specification can be referred to by their GUID, this trait is a building block to interface them in uefi-rs.

You should never need to use the Identify trait directly, but instead go for more specific traits such as Protocol or FileProtocolInfo, which indicate in which circumstances an Identify-tagged type should be used.

For the common case of implementing this trait for a protocol, use the unsafe_protocol macro.

Safety

Implementing Identify is unsafe because attaching an incorrect GUID to a type can lead to type unsafety on both the Rust and UEFI side.

Required Associated Constants

const GUID: Guid

Unique protocol identifier.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".