uefi

Macro cstr16

macro_rules! cstr16 {
    () => { ... };
    ($s:literal) => { ... };
}

Encode a string literal as a &CStr16.

The encoding is done at compile time, so the result can be used in a const item.

An empty string containing just a null character can be created with either cstr16!() or cstr16!("").

Example

use uefi::{CStr16, cstr16};

const S: &CStr16 = cstr16!("abc");
assert_eq!(S.to_u16_slice_with_nul(), [97, 98, 99, 0]);

const EMPTY: &CStr16 = cstr16!();
assert_eq!(EMPTY.to_u16_slice_with_nul(), [0]);
assert_eq!(cstr16!(""), EMPTY);