uefi::mem

Module memory_map

Bundles all relevant types and helpers to work with the UEFI memory map.

To work with the memory map, you should use one of the structs MemoryMapOwned, MemoryMapRef, or MemoryMapRefMut - depending on your use-case. The traits MemoryMap and MemoryMapMut mainly exist to guarantee a streamlined API across these types. We recommend to work with the specific implementation.

Usecase: Obtain UEFI Memory Map

You can use boot::exit_boot_services or boot::memory_map, which returns an properly initialized MemoryMapOwned.

Usecase: Parse Memory Slice as UEFI Memory Map

If you have a chunk of memory and want to parse it as UEFI memory map, which might be the case if a bootloader such as GRUB or Limine passes its boot information, you can use MemoryMapRef or MemoryMapRefMut.

All relevant exports:

the traits MemoryMap and MemoryMapMut,

the trait implementations MemoryMapOwned, MemoryMapRef, and MemoryMapRefMut,

the iterator MemoryMapIter

various associated helper types, such as MemoryMapKey and MemoryMapMeta,

re-exports MemoryDescriptor, MemoryType, and MemoryAttribute.

Structs

MemoryAttribute

Flags describing the capabilities of a memory range.

MemoryDescriptor

A structure describing a region of memory. This type corresponds to version of this struct in the UEFI spec and is always bound to a corresponding UEFI memory map.

MemoryMapIter

An iterator for MemoryMap.

MemoryMapKey

A unique identifier of a UEFI memory map, used to tell the firmware that one has the latest valid memory map when exiting boot services.

MemoryMapMeta

A structure containing the meta attributes associated with a call to GetMemoryMap of UEFI. Note that all values refer to the time this was called. All following invocations (hidden, subtle, and asynchronous ones) will likely invalidate this.

MemoryMapOwned

Implementation of MemoryMapMut that owns the buffer on the UEFI heap.

MemoryMapRef

Implementation of MemoryMap for the given buffer.

MemoryMapRefMut

Implementation of MemoryMapMut for the given buffer.

MemoryType

The type of a memory range.

Enums

MemoryMapError

Errors that may happen when constructing a MemoryMapRef or MemoryMapRefMut.

Traits

MemoryMap

An accessory to the UEFI memory map and associated metadata that can be either iterated or indexed like an array.

MemoryMapMut

Extension to MemoryMap that adds mutable operations. This also includes the ability to sort the memory map.