uefi::mem::memory_map

Trait MemoryMap

pub trait MemoryMap: Debug + Index<usize, Output = MemoryDescriptor> {
    // Required methods
    fn meta(&self) -> MemoryMapMeta;
    fn key(&self) -> MemoryMapKey;
    fn len(&self) -> usize;
    fn buffer(&self) -> &[u8];
    fn entries(&self) -> MemoryMapIter<'_> ;

    // Provided methods
    fn is_empty(&self) -> bool { ... }
    fn get(&self, index: usize) -> Option<&MemoryDescriptor> { ... }
    fn is_sorted(&self) -> bool { ... }
}

An accessory to the UEFI memory map and associated metadata that can be either iterated or indexed like an array.

A MemoryMap is always associated with the unique MemoryMapKey bundled with the map.

To iterate over the entries, call MemoryMap::entries.

UEFI pitfalls

Note that a MemoryMap can quickly become outdated, as soon as any explicit or hidden allocation happens.

As soon as boot services are excited, all previous obtained memory maps must be considered as outdated, except if the MemoryMapKey equals the one returned by exit_boot_services().

Please note that when working with memory maps, the entry_size is usually larger than size_of::<MemoryDescriptor [0]. So to be safe, always use entry_size as step-size when interfacing with the memory map on a low level.

Required Methods

fn meta(&self) -> MemoryMapMeta

Returns the associated MemoryMapMeta.

fn key(&self) -> MemoryMapKey

Returns the associated MemoryMapKey. Note that this isn’t necessarily the key of the latest valid UEFI memory map.

fn len(&self) -> usize

Returns the number of keys in the map.

fn buffer(&self) -> &[u8]

Returns a reference to the underlying memory.

fn entries(&self) -> MemoryMapIter<'_>

Returns an Iterator of type MemoryMapIter.

Provided Methods

fn is_empty(&self) -> bool

Returns if the memory map is empty.

fn get(&self, index: usize) -> Option<&MemoryDescriptor>

Returns a reference to the MemoryDescriptor at the given index, if present.

fn is_sorted(&self) -> bool

Returns if the underlying memory map is sorted regarding the physical address start.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".