uefi::mem::memory_map

Trait MemoryMapMut

pub trait MemoryMapMut: MemoryMap + IndexMut<usize> {
    // Required methods
    fn sort(&mut self);
    unsafe fn buffer_mut(&mut self) -> &mut [u8];

    // Provided method
    fn get_mut(&mut self, index: usize) -> Option<&mut MemoryDescriptor> { ... }
}

Extension to MemoryMap that adds mutable operations. This also includes the ability to sort the memory map.

Required Methods

fn sort(&mut self)

Sorts the memory map by physical address in place. This operation is optional and should be invoked only once.

unsafe fn buffer_mut(&mut self) -> &mut [u8]

Returns a reference to the underlying memory.

Safety

This is unsafe as there is a potential to create invalid entries.

Provided Methods

fn get_mut(&mut self, index: usize) -> Option<&mut MemoryDescriptor>

Returns a mutable reference to the MemoryDescriptor at the given index, if present.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".