uefi

Module prelude

This module is used to simplify importing the most common UEFI types.

This includes the system table modules, Status codes, etc.

Re-exports

pub use crate::boot;

pub use crate::runtime;

pub use crate::system;

pub use crate::Handle;

pub use crate::ResultExt;

pub use crate::StatusExt;

Macros

cstr8

Encode a string literal as a &CStr8.

cstr16

Encode a string literal as a &CStr16.

Structs

Status

UEFI uses status codes in order to report successes, errors, and warnings.

Attribute Macros

entry

Custom attribute for a UEFI executable entry point.