Struct uefi::prelude::Status

#[must_use]

#[repr(transparent)]
pub struct Status(pub usize);

UEFI uses status codes in order to report successes, errors, and warnings.

Unfortunately, the spec allows and encourages implementation-specific non-portable status codes. Therefore, these cannot be modeled as a Rust enum, as injecting an unknown value in a Rust enum is undefined behaviour.

For lack of a better option, we therefore model them as a newtype of usize.

Implementations

impl Status

pub const SUCCESS: Status

The operation completed successfully.

pub const WARN_UNKNOWN_GLYPH: Status

The string contained characters that could not be rendered and were skipped.

pub const WARN_DELETE_FAILURE: Status

The handle was closed, but the file was not deleted.

pub const WARN_WRITE_FAILURE: Status

The handle was closed, but the data to the file was not flushed properly.

pub const WARN_BUFFER_TOO_SMALL: Status

The resulting buffer was too small, and the data was truncated.

pub const WARN_STALE_DATA: Status

The data has not been updated within the timeframe set by local policy.

pub const WARN_FILE_SYSTEM: Status

The resulting buffer contains UEFI-compliant file system.

pub const WARN_RESET_REQUIRED: Status

The operation will be processed across a system reset.

pub const LOAD_ERROR: Status

The image failed to load.

pub const INVALID_PARAMETER: Status

A parameter was incorrect.

pub const UNSUPPORTED: Status

The operation is not supported.

pub const BAD_BUFFER_SIZE: Status

The buffer was not the proper size for the request.

pub const BUFFER_TOO_SMALL: Status

The buffer is not large enough to hold the requested data. The required buffer size is returned in the appropriate parameter.

pub const NOT_READY: Status

There is no data pending upon return.

pub const DEVICE_ERROR: Status

The physical device reported an error while attempting the operation.

pub const WRITE_PROTECTED: Status

The device cannot be written to.

pub const OUT_OF_RESOURCES: Status

A resource has run out.

pub const VOLUME_CORRUPTED: Status

An inconstency was detected on the file system.

pub const VOLUME_FULL: Status

There is no more space on the file system.

pub const NO_MEDIA: Status

The device does not contain any medium to perform the operation.

pub const MEDIA_CHANGED: Status

The medium in the device has changed since the last access.

pub const NOT_FOUND: Status

The item was not found.

pub const ACCESS_DENIED: Status

Access was denied.

pub const NO_RESPONSE: Status

The server was not found or did not respond to the request.

pub const NO_MAPPING: Status

A mapping to a device does not exist.

pub const TIMEOUT: Status

The timeout time expired.

pub const NOT_STARTED: Status

The protocol has not been started.

pub const ALREADY_STARTED: Status

The protocol has already been started.

pub const ABORTED: Status

The operation was aborted.

pub const ICMP_ERROR: Status

An ICMP error occurred during the network operation.

pub const TFTP_ERROR: Status

A TFTP error occurred during the network operation.

pub const PROTOCOL_ERROR: Status

A protocol error occurred during the network operation.

pub const INCOMPATIBLE_VERSION: Status

The function encountered an internal version that was incompatible with a version requested by the caller.

pub const SECURITY_VIOLATION: Status

The function was not performed due to a security violation.

pub const CRC_ERROR: Status

A CRC error was detected.

pub const END_OF_MEDIA: Status

Beginning or end of media was reached

pub const END_OF_FILE: Status

The end of the file was reached.

pub const INVALID_LANGUAGE: Status

The language specified was invalid.

pub const COMPROMISED_DATA: Status

The security status of the data is unknown or compromised and the data must be updated or replaced to restore a valid security status.

pub const IP_ADDRESS_CONFLICT: Status

There is an address conflict address allocation

pub const HTTP_ERROR: Status

A HTTP error occurred during the network operation.

impl Status

pub fn is_success(self) -> bool

Returns true if status code indicates success.

pub fn is_warning(self) -> bool

Returns true if status code indicates a warning.

pub fn is_error(self) -> bool

Returns true if the status code indicates an error.

pub fn into_with_val<T>(self, val: impl FnOnce() -> T) -> Result<T, ()>

Converts this status code into a result with a given value.

pub fn into_with_err<ErrData: Debug>(
self,
err: impl FnOnce(Status) -> ErrData
) -> Result<(), ErrData>

Converts this status code into a result with a given error payload

pub fn into_with<T, ErrData: Debug>(
self,
val: impl FnOnce() -> T,
err: impl FnOnce(Status) -> ErrData
) -> Result<T, ErrData>

Convert this status code into a result with a given value and error payload