uefi

Module proto

Protocol definitions.

Protocols are sets of related functionality identified by a unique ID. They can be implemented by a UEFI driver or occasionally by a UEFI application.

See the boot documentation for details of how to open a protocol.

Modules

console

Console support protocols.

debug

Provides support for the UEFI debugging protocol.

device_path

Device Path protocol

driver

UEFI driver model protocols.

loaded_image

LoadedImage protocol.

media

Media access protocols.

misc

Miscellaneous protocols.

network

Network access protocols.

pi

Platform Initialization protocols.

rng

Rng protocol.

security

Protocols related to secure technologies.

shell_params

ShellParams protocol

shim

Shim lock protocol.

string

String protocols.

tcg

TCG (Trusted Computing Group) protocols.

Enums

BootPolicy

The UEFI boot policy is a property that influences the behaviour of various UEFI functions that load files (typically UEFI images).

BootPolicyError

Errors that can happen when working with BootPolicy.

Traits

Protocol

Common trait implemented by all standard UEFI protocols.

ProtocolPointer

Trait for creating a protocol pointer from a c_void pointer.

Attribute Macros

unsafe_protocol

Attribute macro for marking structs as UEFI protocols.