Module uefi::proto::console

Console support protocols.

The console represents the various input and output methods used by the user to interact with the early boot platform.

Modules

gop
Graphics output protocol.

pointer
Pointer device access.

serial
Abstraction over byte stream devices, also known as serial I/O devices.

text
Text I/O.