Enum uefi::proto::console::gop::BltOp

pub enum BltOp<'buf> {
    VideoFill {
        color: BltPixel,
        dest: (usize, usize),
        dims: (usize, usize),
    },
    VideoToBltBuffer {
        buffer: &'buf mut [BltPixel],
        src: (usize, usize),
        dest: BltRegion,
        dims: (usize, usize),
    },
    BufferToVideo {
        buffer: &'buf [BltPixel],
        src: BltRegion,
        dest: (usize, usize),
        dims: (usize, usize),
    },
    VideoToVideo {
        src: (usize, usize),
        dest: (usize, usize),
        dims: (usize, usize),
    },
}

Blit operation to perform.

Variants

VideoFill

Fills a rectangle of video display with a pixel color.

Fields of VideoFill

color: BltPixel

The color to fill with.

dest: (usize, usize)

The X / Y coordinates of the destination rectangle.

dims: (usize, usize)

The width / height of the rectangle.

VideoToBltBuffer

Reads data from the video display to the buffer.

Fields of VideoToBltBuffer

buffer: &'buf mut [BltPixel]

Buffer into which to copy data.

src: (usize, usize)

Coordinates of the source rectangle, in the frame buffer.

dest: BltRegion

Location of the destination rectangle in the user-provided buffer

dims: (usize, usize)

Width / height of the rectangles.

BufferToVideo

Write data from the buffer to the video rectangle. Delta must be the stride (count of bytes in a row) of the buffer.

Fields of BufferToVideo

buffer: &'buf [BltPixel]

Buffer from which to copy data.

src: BltRegion

Location of the source rectangle in the user-provided buffer.

dest: (usize, usize)

Coordinates of the destination rectangle, in the frame buffer.

dims: (usize, usize)

Width / height of the rectangles.

VideoToVideo

Copy from the source rectangle in video memory to the destination rectangle, also in video memory.

Fields of VideoToVideo

src: (usize, usize)

Coordinates of the source rectangle, in the frame buffer.

dest: (usize, usize)

Coordinates of the destination rectangle, also in the frame buffer.

dims: (usize, usize)

Width / height of the rectangles.