Struct uefi::proto::console::pointer::Pointer

#[repr(C)]pub struct Pointer<'boot> { /* fields omitted */ }

Provides information about a pointer device.

Implementations

impl<'boot> Pointer<'boot>

pub fn reset(&mut self, extended_verification: bool) -> Result

Resets the pointer device hardware.

The extended_verification parameter is used to request that UEFI performs an extended check and reset of the input device.

Errors

DeviceError if the device is malfunctioning and cannot be reset.

pub fn read_state(&mut self) -> Result<Option<PointerState>>

Retrieves the pointer device’s current state, if a state change occured since the last time this function was called.

Use wait_for_input_event() with the BootServices::wait_for_event() interface in order to wait for input from the pointer device.

Errors

DeviceError if there was an issue with the pointer device.

pub fn wait_for_input_event(&self) -> Event

Event to be used with BootServices::wait_for_event() in order to wait for input from the pointer device

pub fn mode(&self) -> &PointerMode

Returns a reference to the pointer device information.