uefi::proto::console::serial

Struct Serial

pub struct Serial(/* private fields */);

Provides access to a serial I/O device.

This can include standard UART devices, serial ports over a USB interface, or any other character-based communication device.

Since UEFI drivers are implemented through polling, if you fail to regularly check for input/output, some data might be lost.

Implementations

impl Serial

pub fn reset(&mut self) -> Result

Reset the device.

pub const fn io_mode(&self) -> &IoMode

Returns the current I/O mode.

pub fn set_attributes(&mut self, mode: &IoMode) -> Result

Sets the device’s new attributes.

The given IoMode will become the device’s new IoMode, with some exceptions:

control_mask is ignored, since it’s a read-only field;

values set to 0 / Default will be filled with the device’s default parameters

if either baud_rate or receive_fifo_depth is less than the device’s minimum, an error will be returned; this value will be rounded down to the nearest value supported by the device;

pub fn get_control_bits(&self) -> Result<ControlBits>

Retrieve the device’s current control bits.

pub fn set_control_bits(&mut self, bits: ControlBits) -> Result

Sets the device’s new control bits.

Not all bits can be modified with this function. A mask of the allowed bits is stored in the ControlBits::SETTABLE constant.

pub fn read(&mut self, data: &mut [u8]) -> Result<(), usize>

Reads data from this device.

This operation will block until the buffer has been filled with data or an error occurs. In the latter case, the error will indicate how many bytes were actually read from the device.

pub fn write(&mut self, data: &[u8]) -> Result<(), usize>

Writes data to this device.

This operation will block until the data has been fully written or an error occurs. In the latter case, the error will indicate how many bytes were actually written to the device.