Struct uefi::proto::console::text::Input

#[repr(C)]
pub struct Input { /* fields omitted */ }

Interface for text-based input devices.

Implementations

impl Input

pub fn reset(&mut self, extended_verification: bool) -> Result

Resets the input device hardware.

The extended_verification parameter is used to request that UEFI performs an extended check and reset of the input device.

Errors

DeviceError if the device is malfunctioning and cannot be reset.

pub fn read_key(&mut self) -> Result<Option<Key>>

Reads the next keystroke from the input device, if any.

Use wait_for_key_event() with the BootServices::wait_for_event() interface in order to wait for a key to be pressed.

Errors

DeviceError if there was an issue with the input device

pub fn wait_for_key_event(&self) -> Event

Event to be used with BootServices::wait_for_event() in order to wait for a key to be available