Struct uefi::proto::console::text::Output

#[repr(C)]
pub struct Output<'boot> { /* fields omitted */ }

Interface for text-based output devices.

It implements the fmt::Write trait, so you can use it to print text with standard Rust constructs like the write!() and writeln!() macros.

Implementations

impl<'boot> Output<'boot>

pub fn reset(&mut self, extended: bool) -> Result

Resets and clears the text output device hardware.

pub fn clear(&mut self) -> Result

Clears the output screen.

The background is set to the current background color. The cursor is moved to (0, 0).

pub fn output_string(&mut self, string: &CStr16) -> Result

Writes a string to the output device.

pub fn test_string(&mut self, string: &CStr16) -> Result<bool>

Checks if a string contains only supported characters.

UEFI applications are encouraged to try to print a string even if it contains some unsupported characters.

pub fn modes<'out>(&'out mut self) -> OutputModeIter<'out, 'boot>

Returns an iterator of all supported text modes.

pub fn current_mode(&self) -> Result<Option<OutputMode>>

Returns the the current text mode.

pub fn set_mode(&mut self, mode: OutputMode) -> Result

Sets a mode as current.

pub fn cursor_visible(&self) -> bool

Returns whether the cursor is currently shown or not.

pub fn enable_cursor(&mut self, visible: bool) -> Result

Make the cursor visible or invisible.

The output device may not support this operation, in which case an Unsupported error will be returned.

pub fn cursor_position(&self) -> (usize, usize)

Returns the column and row of the cursor.

pub fn set_cursor_position(&mut self, column: usize, row: usize) -> Result

Sets the cursor’s position, relative to the top-left corner, which is (0, 0).

This function will fail if the cursor’s new position would exceed the screen’s bounds.

pub fn set_color(&mut self, foreground: Color, background: Color) -> Result

Sets the text and background colors for the console.

Note that for the foreground color you can choose any color. The background must be one of the first 8 colors.