uefi::proto

Module debug

Provides support for the UEFI debugging protocol.

This protocol is designed to allow debuggers to query the state of the firmware, as well as set up callbacks for various events.

It also defines a Debugport protocol for debugging over serial devices.

An example UEFI debugger is Intel’s UDK Debugger Tool.

Structs

DebugPort

The debug port protocol abstracts the underlying debug port hardware, whether it is a regular Serial port or something else.

DebugSupport

The debugging support protocol allows debuggers to connect to a UEFI machine. It is expected that there will typically be two instances of the EFI Debug Support protocol in the system. One associated with the native processor instruction set (IA-32, x64, ARM, RISC-V, or Itanium processor family), and one for the EFI virtual machine that implements EFI byte code (EBC). While multiple instances of the EFI Debug Support protocol are expected, there must never be more than one for any given instruction set.

ExceptionType

Represents supported CPU exceptions.

ProcessorArch

The instruction set architecture of the running processor.

Unions

SystemContext

Universal EFI_SYSTEM_CONTEXT definition This is passed to debug callbacks