uefi::proto::device_path::acpi

Struct Expanded

pub struct Expanded { /* private fields */ }

Expanded ACPI device path node.

Implementations

impl Expanded

pub fn hid(&self) -> u32

Device’s PnP hardware ID stored in a numeric 32-bit compressed EISA-type ID.

pub fn uid(&self) -> u32

Unique ID that is required by ACPI if two devices have the same HID.

pub fn cid(&self) -> u32

Device’s compatible PnP hardware ID stored in a numeric 32-bit compressed EISA-type ID.

pub fn hid_str(&self) -> &[u8]

Device’s PnP hardware ID stored as a null-terminated ASCII string. This value must match the corresponding HID in the ACPI name space. If the length of this string not including the null-terminator is 0, then the numeric HID is used.

pub fn uid_str(&self) -> &[u8]

Unique ID that is required by ACPI if two devices have the same HID. This value is stored as a null-terminated ASCII string. If the length of this string not including the null-terminator is 0, then the numeric UID is used.

pub fn cid_str(&self) -> &[u8]

Device’s compatible PnP hardware ID stored as a null-terminated ASCII string. If the length of this string not including the null-terminator is 0, then the numeric CID is used.