uefi::proto::device_path

Module build

Utilities for creating new DevicePaths.

This module contains DevicePathBuilder, as well as submodules containing types for building each type of device path node.

Modules

acpi

Device path build nodes for DeviceType::ACPI.

bios_boot_spec

Device path build nodes for DeviceType::BIOS_BOOT_SPEC.

end

Device path build nodes for DeviceType::END.

hardware

Device path build nodes for DeviceType::HARDWARE.

media

Device path build nodes for DeviceType::MEDIA.

messaging

Device path build nodes for DeviceType::MESSAGING.

Structs

DevicePathBuilder

A builder for DevicePaths.

Enums

BuildError

Error type used by DevicePathBuilder.

Traits

BuildNode

Trait for types that can be used to build a node via DevicePathBuilder::push.