uefi::proto::device_path::build

Trait BuildNode

pub unsafe trait BuildNode {
    // Required methods
    fn size_in_bytes(&self) -> Result<u16, BuildError>;
    fn write_data(&self, out: &mut [MaybeUninit<u8>]);
}

Trait for types that can be used to build a node via DevicePathBuilder::push.

This trait is implemented for all the node types in uefi::proto::device_path::build. It is also implemented for &DevicePathNode, which allows an existing node to be copied by the builder.

Safety

The methods of this trait are safe to call, but the trait itself is unsafe because an incorrect implementation could cause unsafety. In particular, the write_data is required to completely initialize all bytes in the output slice.

Required Methods

fn size_in_bytes(&self) -> Result<u16, BuildError>

Size of the node in bytes, including the standard node header. Returns BuildError::NodeTooBig if the node’s size does not fit in a u16.

fn write_data(&self, out: &mut [MaybeUninit<u8>])

Write out the node data.

The length of out must be equal to the node’s size_in_bytes.

The out slice will be fully initialized after the call.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".