uefi::proto::device_path

Struct DevicePathNode

pub struct DevicePathNode { /* private fields */ }

A single node within a DevicePath.

Each node starts with a DevicePathHeader. The rest of the data in the node depends on the type of node. You can “cast” a node to a specific one like this:

use uefi::proto::device_path::DevicePath;
use uefi::proto::device_path::media::FilePath;

let image_device_path: &DevicePath = unsafe { DevicePath::from_ffi_ptr(0x1337 as *const _) };
let file_path = image_device_path
        .node_iter()
        .find_map(|node| {
            let node: &FilePath = node.try_into().ok()?;
            let path = node.path_name().to_cstring16().ok()?;
            Some(path.to_string().to_uppercase())
        });

More types are available in uefi::proto::device_path. Builder types can be found in uefi::proto::device_path::build

See the module-level documentation for more details.

Implementations

impl DevicePathNode

pub unsafe fn from_ffi_ptr<'a>(ptr: *const FfiDevicePath) -> &'a Self

Create a DevicePathNode reference from an opaque pointer.

Safety

The input pointer must point to valid data. That data must remain valid for the lifetime 'a, and cannot be mutated during that lifetime.

pub const fn as_ffi_ptr(&self) -> *const FfiDevicePath

Cast to a FfiDevicePath pointer.

pub const fn device_type(&self) -> DeviceType

Type of device

pub const fn sub_type(&self) -> DeviceSubType

Sub type of device

pub const fn full_type(&self) -> (DeviceType, DeviceSubType)

Tuple of the node’s type and subtype.

pub const fn length(&self) -> u16

Size (in bytes) of the full DevicePathNode, including the header.

pub fn is_end_entire(&self) -> bool

True if this node ends an entire DevicePath.

pub const fn data(&self) -> &[u8]

Returns the payload data of this node.

pub fn as_enum(&self) -> Result<DevicePathNodeEnum<'_>, NodeConversionError>

Convert from a generic DevicePathNode reference to an enum of more specific node types.