uefi::proto::device_path

Struct LoadedImageDevicePath

pub struct LoadedImageDevicePath(/* private fields */);

Protocol for accessing the device path that was passed in to load_image when loading a PE/COFF image.

The layout of this type is the same as a DevicePath.