uefi::proto

Module driver

UEFI driver model protocols.

Structs

ComponentName1Deprecated

Protocol that provides human-readable names for a driver and for each of the controllers that the driver is managing.

ComponentName2

Protocol that provides human-readable names for a driver and for each of the controllers that the driver is managing.

LanguageIter

Iterator returned by ComponentName1::supported_languages and ComponentName2::supported_languages.

Enums

ComponentName

Wrapper around ComponentName1 and ComponentName2. This will use ComponentName2 if available, otherwise it will back to ComponentName1.

LanguageError

Error returned by ComponentName1::supported_languages and ComponentName2::supported_languages.