Struct uefi::proto::loaded_image::LoadedImage

#[repr(C)]
pub struct LoadedImage { /* fields omitted */ }

The LoadedImage protocol. This can be opened on any image handle using the HandleProtocol boot service.

Implementations

impl LoadedImage

pub fn device(&self) -> Handle

Returns a handle to the storage device on which the image is located.

pub fn load_options<'a>(
&self,
buffer: &'a mut [u8]
) -> Result<&'a str, LoadOptionsError>

Get the load options of the given image. If the image was executed from the EFI shell, or from a boot option, this is the command line that was used to execute it as a string. If no options were given, this returns Ok("").

pub fn info(&self) -> (usize, u64)

Returns the base address and the size in bytes of the loaded image.