Struct uefi::proto::media::block::BlockIOMedia

#[repr(C)]
pub struct BlockIOMedia { /* fields omitted */ }

Media information structure

Implementations

impl BlockIOMedia

pub fn media_id(&self) -> u32

The current media ID.

pub fn is_removable_media(&self) -> bool

True if the media is removable.

pub fn is_media_preset(&self) -> bool

True if there is a media currently present in the device.

pub fn is_logical_partition(&self) -> bool

True if block IO was produced to abstract partition structure.

pub fn is_read_only(&self) -> bool

True if the media is marked read-only.

pub fn is_write_caching(&self) -> bool

True if writeBlocks function writes data.

pub fn block_size(&self) -> u32

The intrinsic block size of the device.

If the media changes, then this field is updated. Returns the number of bytes per logical block.

pub fn io_align(&self) -> u32

Supplies the alignment requirement for any buffer used in a data transfer.

pub fn last_block(&self) -> Lba

The last LBA on the device. If the media changes, then this field is updated.

pub fn lowest_aligned_lba(&self) -> Lba

Returns the first LBA that is aligned to a physical block boundary.

pub fn logical_blocks_per_physical_block(&self) -> u32

Returns the number of logical blocks per physical block.

pub fn optimal_transfer_length_granularity(&self) -> u32

Returns the optimal transfer length granularity as a number of logical blocks.