uefi::proto::media::disk

Struct DiskIo

pub struct DiskIo(/* private fields */);

The disk I/O protocol.

This protocol is used to abstract the block accesses of the block I/O protocol to a more general offset-length protocol. Firmware is responsible for adding this protocol to any block I/O interface that appears in the system that does not already have a disk I/O protocol.

Implementations

impl DiskIo

pub fn read_disk(&self, media_id: u32, offset: u64, buffer: &mut [u8]) -> Result

Reads bytes from the disk device.

Arguments:

media_id - ID of the medium to be read.

offset - Starting byte offset on the logical block I/O device to read from.

buffer - Pointer to a buffer to read into.

Errors:

uefi::status::INVALID_PARAMETER The read request contains device addresses that are not valid for the device.

uefi::status::DEVICE_ERROR The device reported an error while performing the read operation.

uefi::status::NO_MEDIA There is no medium in the device.

uefi::status::MEDIA_CHANGED media_id is not for the current medium.

pub fn write_disk( &mut self, media_id: u32, offset: u64, buffer: &[u8], ) -> Result

Writes bytes to the disk device.

Arguments:

media_id - ID of the medium to be written.

offset - Starting byte offset on the logical block I/O device to write to.

buffer - Pointer to a buffer to write from.

Errors:

uefi::status::INVALID_PARAMETER The write request contains device addresses that are not valid for the device.

uefi::status::DEVICE_ERROR The device reported an error while performing the write operation.

uefi::status::NO_MEDIA There is no medium in the device.

uefi::status::MEDIA_CHANGED media_id is not for the current medium.

uefi::status::WRITE_PROTECTED The device cannot be written to.