uefi::proto::media::file

Struct FileSystemInfo

pub struct FileSystemInfo { /* private fields */ }

System volume information

May only be obtained on the root directory’s file handle.

Please note that only the system volume’s volume label may be set using this information structure. Consider using FileSystemVolumeLabel instead.

Implementations

impl FileSystemInfo

pub fn new<'buf>( storage: &'buf mut [u8], read_only: bool, volume_size: u64, free_space: u64, block_size: u32, volume_label: &CStr16, ) -> Result<&'buf mut Self, FileInfoCreationError>

Create a FileSystemInfo structure

The structure will be created in-place within the provided storage buffer. The buffer must be large enough to hold the data structure, including a null-terminated UCS-2 name string.

The buffer must be correctly aligned. You can query the required alignment using the alignment() method of the Align trait that this struct implements.

pub const fn read_only(&self) -> bool

Truth that the volume only supports read access

pub const fn volume_size(&self) -> u64

Number of bytes managed by the file system

pub const fn free_space(&self) -> u64

Number of available bytes for use by the file system

pub const fn block_size(&self) -> u32

Nominal block size by which files are typically grown

pub fn volume_label(&self) -> &CStr16

Volume label