uefi::proto::media::file

Struct RegularFile

pub struct RegularFile(/* private fields */);

A FileHandle that is also a regular (data) file.

Use FileHandle::into_type or RegularFile::new to create a RegularFile. In addition to supporting the normal File operations, RegularFile supports direct reading and writing.

Implementations

impl RegularFile

pub const END_OF_FILE: u64 = u64::MAX

A special position used to seek to the end of a file with set_position().

pub const unsafe fn new(handle: FileHandle) -> Self

Coverts a FileHandle into a RegularFile without checking the file kind.

Safety

This function should only be called on handles which ARE NOT directories, doing otherwise is unsafe.

pub fn read(&mut self, buffer: &mut [u8]) -> Result<usize>

Read data from file.

Try to read as much as possible into buffer. Returns the number of bytes that were actually read.

Arguments

buffer The target buffer of the read operation

Errors

See section EFI_FILE_PROTOCOL.Read() in the UEFI Specification for more details.

uefi::Status::NO_MEDIA

uefi::Status::DEVICE_ERROR

uefi::Status::VOLUME_CORRUPTED

Quirks

Some UEFI implementations have a bug where large reads will incorrectly return an error. This function avoids that bug by reading in chunks of no more than 1 MiB. This is handled internally within the function; callers can safely pass in a buffer of any size. See https://github.com/rust-osdev/uefi-rs/issues/825 for more information.

pub fn write(&mut self, buffer: &[u8]) -> Result<(), usize>

Write data to file

Write buffer to file, increment the file pointer.

If an error occurs, returns the number of bytes that were actually written. If no error occurred, the entire buffer is guaranteed to have been written successfully.

Arguments

buffer Buffer to write to file

Errors

See section EFI_FILE_PROTOCOL.Write() in the UEFI Specification for more details.

uefi::Status::NO_MEDIA

uefi::Status::DEVICE_ERROR

uefi::Status::VOLUME_CORRUPTED

uefi::Status::WRITE_PROTECTED

uefi::Status::ACCESS_DENIED

uefi::Status::VOLUME_FULL

pub fn get_position(&mut self) -> Result<u64>

Get the file’s current position

Errors

See section EFI_FILE_PROTOCOL.GetPosition() in the UEFI Specification for more details.

uefi::Status::DEVICE_ERROR

pub fn set_position(&mut self, position: u64) -> Result

Sets the file’s current position

Set the position of this file handle to the absolute position specified by position.

Seeking past the end of the file is allowed, it will trigger file growth on the next write. Using a position of RegularFile::END_OF_FILE will seek to the end of the file.

Arguments

position The new absolution position of the file handle

Errors

See section EFI_FILE_PROTOCOL.SetPosition() in the UEFI Specification for more details.

uefi::Status::DEVICE_ERROR