Type Definition uefi::proto::media::file::FileSystemVolumeLabel

type FileSystemVolumeLabel = NamedFileProtocolInfo<FileSystemVolumeLabelHeader>;

System volume label

May only be obtained on the root directory’s file handle.

Implementations

impl FileSystemVolumeLabel

pub fn new<'buf>(
storage: &'buf mut [u8],
volume_label: &str
) -> Result<&'buf mut Self, FileInfoCreationError>

Create a FileSystemVolumeLabel structure

The structure will be created in-place within the provided storage buffer. The buffer must be large enough to hold the data structure, including a null-terminated UCS-2 version of the name string.

The buffer must be correctly aligned. You can query the required alignment using the alignment() method of the Align trait that this struct implements.

pub fn volume_label(&self) -> &CStr16

Volume label