uefi::proto::network

Module pxe

PXE Base Code protocol.

Structs

ArpEntry

An entry for the ARP cache found in Mode::arp_cache

BaseCode

PXE Base Code protocol

DhcpV4Flags

Represents the ‘flags’ field for a DhcpV4Packet.

DhcpV4Packet

A Dhcpv4 Packet.

DhcpV6Packet

A Dhcpv6 Packet.

DiscoverInfo

This struct contains optional parameters for BaseCode::discover.

FfiDiscoverInfo

Opaque type that should be used to represent a pointer to a DiscoverInfo in foreign function interfaces. This type produces a thin pointer, unlike DiscoverInfo.

IcmpError

An ICMP error packet.

IcmpErrorEcho

Corresponds to the Echo field in the anonymous union inside EFI_PXE_BASE_CODE_ICMP_ERROR in the C API.

IpFilter

IP receive filter settings

IpFilters

IP receive filters.

Mode

The data values in this structure are read-only and are updated by the BaseCode.

MtftpFileInfo

Returned by BaseCode::mtftp_read_dir.

MtftpInfo

MTFTP connection parameters

ReadDirParseError

Returned if a server sends a malformed response in BaseCode::tftp_read_dir or BaseCode::mtftp_read_dir.

RouteEntry

An entry for the route table found in Mode::route_table

Server

An entry in the Boot Server list

TftpError

A TFTP error packet.

TftpFileInfo

Returned by BaseCode::tftp_read_dir.

UdpOpFlags

Flags for UDP read and write operations.

Enums

BootstrapType

A type of bootstrap to perform in BaseCode::discover.

Unions

IcmpErrorUnion

Corresponds to the anonymous union inside EFI_PXE_BASE_CODE_ICMP_ERROR in the C API.

Packet

A network packet.