uefi::proto::network::pxe

Struct DhcpV4Packet

pub struct DhcpV4Packet {
    pub bootp_opcode: u8,
    pub bootp_hw_type: u8,
    pub bootp_hw_addr_len: u8,
    pub bootp_gate_hops: u8,
    pub bootp_ci_addr: [u8; 4],
    pub bootp_yi_addr: [u8; 4],
    pub bootp_si_addr: [u8; 4],
    pub bootp_gi_addr: [u8; 4],
    pub bootp_hw_addr: [u8; 16],
    pub bootp_srv_name: [u8; 64],
    pub bootp_boot_file: [u8; 128],
    pub dhcp_options: [u8; 56],
    /* private fields */
}

A Dhcpv4 Packet.

Corresponds to the EFI_PXE_BASE_CODE_DHCPV4_PACKET type in the C API.

Fields
bootp_opcode: u8

Packet op code / message type.

bootp_hw_type: u8

Hardware address type.

bootp_hw_addr_len: u8

Hardware address length.

bootp_gate_hops: u8

Client sets to zero, optionally used by gateways in cross-gateway booting.

bootp_ci_addr: [u8; 4]

Client IP address, filled in by client in bootrequest if known.

bootp_yi_addr: [u8; 4]

‘your’ (client) IP address; filled by server if client doesn’t know its own address (bootp_ci_addr was 0).

bootp_si_addr: [u8; 4]

Server IP address, returned in bootreply by server.

bootp_gi_addr: [u8; 4]

Gateway IP address, used in optional cross-gateway booting.

bootp_hw_addr: [u8; 16]

Client hardware address, filled in by client.

bootp_srv_name: [u8; 64]

Optional server host name, null terminated string.

bootp_boot_file: [u8; 128]

Boot file name, null terminated string, ‘generic’ name or null in bootrequest, fully qualified directory-path name in bootreply.

dhcp_options: [u8; 56]

Optional vendor-specific area, e.g. could be hardware type/serial on request, or ‘capability’ / remote file system handle on reply. This info may be set aside for use by a third phase bootstrap or kernel.

Implementations

impl DhcpV4Packet

pub const DHCP_MAGIK: u32 = 0x63825363

The expected value for Self::dhcp_magik.

pub const fn bootp_ident(&self) -> u32

Transaction ID, a random number, used to match this boot request with the responses it generates.

pub const fn bootp_seconds(&self) -> u16

Filled in by client, seconds elapsed since client started trying to boot.

pub const fn bootp_flags(&self) -> DhcpV4Flags

The flags.

pub const fn dhcp_magik(&self) -> u32

A magic cookie, should be Self::DHCP_MAGIK.