uefi::proto::network::pxe

Struct DiscoverInfo

pub struct DiscoverInfo { /* private fields */ }

This struct contains optional parameters for BaseCode::discover.

Corresponds to the EFI_PXE_BASE_CODE_DISCOVER_INFO type in the C API.

Implementations

impl DiscoverInfo

pub fn new_in_buffer<'buf>( buffer: &'buf mut [MaybeUninit<u8>], use_m_cast: bool, use_b_cast: bool, use_u_cast: bool, must_use_list: bool, server_m_cast_ip: IpAddress, srv_list: &[Server], ) -> Result<&'buf mut Self>

Create a DiscoverInfo.

impl DiscoverInfo

pub const fn use_m_cast(&self) -> bool

Returns whether discovery should use multicast.

pub const fn use_b_cast(&self) -> bool

Returns whether discovery should use broadcast.

pub const fn use_u_cast(&self) -> bool

Returns whether discovery should use unicast.

pub const fn must_use_list(&self) -> bool

Returns whether discovery should only accept boot servers in the server list (boot server verification).

pub const fn server_m_cast_ip(&self) -> &IpAddress

Returns the address used in multicast discovery.

pub const fn ip_cnt(&self) -> u16

Returns the amount of Boot Server.

pub const fn srv_list(&self) -> &[Server]

Returns the Boot Server list used for unicast discovery or boot server verification.