uefi::proto::network::pxe

Struct Mode

#[repr(C)]
pub struct Mode {

    pub started: bool,
    pub ipv6_available: bool,
    pub ipv6_supported: bool,
    pub using_ipv6: bool,
    pub bis_supported: bool,
    pub bis_detected: bool,
    pub auto_arp: bool,
    pub send_guid: bool,
    pub dhcp_discover_valid: bool,
    pub dhcp_ack_received: bool,
    pub proxy_offer_received: bool,
    pub pxe_discover_valid: bool,
    pub pxe_reply_received: bool,
    pub pxe_bis_reply_received: bool,
    pub icmp_error_received: bool,
    pub tftp_error_received: bool,
    pub make_callbacks: bool,
    pub ttl: u8,
    pub tos: u8,
    pub station_ip: IpAddress,
    pub subnet_mask: IpAddress,
    pub dhcp_discover: Packet,
    pub dhcp_ack: Packet,
    pub proxy_offer: Packet,
    pub pxe_discover: Packet,
    pub pxe_reply: Packet,
    pub pxe_bis_reply: Packet,
    pub ip_filter: IpFilter,
    pub arp_cache_entries: u32,
    pub arp_cache: [ArpEntry; 8],
    pub route_table_entries: u32,
    pub route_table: [RouteEntry; 8],
    pub icmp_error: IcmpError,
    pub tftp_error: TftpError,

}

The data values in this structure are read-only and are updated by the BaseCode.

Corresponds to the EFI_PXE_BASE_CODE_MODE type in the C API.

Fields
started: bool

true if this device has been started by calling BaseCode::start. This field is set to true by BaseCode::start and to false by the BaseCode::stop function.

ipv6_available: bool

true if the UNDI protocol supports IPv6

ipv6_supported: bool

true if this PXE Base Code Protocol implementation supports IPv6.

using_ipv6: bool

true if this device is currently using IPv6. This field is set by BaseCode::start.

bis_supported: bool

true if this PXE Base Code implementation supports Boot Integrity Services (BIS). This field is set by BaseCode::start.

bis_detected: bool

true if this device and the platform support Boot Integrity Services (BIS). This field is set by BaseCode::start.

auto_arp: bool

true for automatic ARP packet generation, false otherwise. This field is initialized to true by BaseCode::start and can be modified with BaseCode::set_parameters.

send_guid: bool

This field is used to change the Client Hardware Address (chaddr) field in the DHCP and Discovery packets. Set to true to send the SystemGuid (if one is available). Set to false to send the client NIC MAC address. This field is initialized to false by BaseCode::start and can be modified with BaseCode::set_parameters.

dhcp_discover_valid: bool

This field is initialized to false by BaseCode::start and set to true when BaseCode::dhcp completes successfully. When true, Self::dhcp_discover is valid. This field can also be changed by BaseCode::set_packets.

dhcp_ack_received: bool

This field is initialized to false by BaseCode::start and set to true when BaseCode::dhcp completes successfully. When true, Self::dhcp_ack is valid. This field can also be changed by BaseCode::set_packets.

proxy_offer_received: bool

This field is initialized to false by BaseCode::start and set to true when BaseCode::dhcp completes successfully and a proxy DHCP offer packet was received. When true, Self::proxy_offer is valid. This field can also be changed by BaseCode::set_packets.

pxe_discover_valid: bool

When true, Self::pxe_discover is valid. This field is set to false by BaseCode::start and BaseCode::dhcp, and can be set to true or false by BaseCode::discover and BaseCode::set_packets.

pxe_reply_received: bool

When true, Self::pxe_reply is valid. This field is set to false by BaseCode::start and BaseCode::dhcp, and can be set to true or false by BaseCode::discover and BaseCode::set_packets.

pxe_bis_reply_received: bool

When true, Self::pxe_bis_reply is valid. This field is set to false by BaseCode::start and BaseCode::dhcp, and can be set to true or false by the BaseCode::discover and BaseCode::set_packets.

icmp_error_received: bool

Indicates whether Self::icmp_error has been updated. This field is reset to false by BaseCode::start, BaseCode::dhcp, BaseCode::discover,BaseCode::udp_read, BaseCode::udp_write, BaseCode::arp and any of the TFTP/MTFTP operations. If an ICMP error is received, this field will be set to true after Self::icmp_error is updated.

tftp_error_received: bool

Indicates whether Self::tftp_error has been updated. This field is reset to false by BaseCode::start and any of the TFTP/MTFTP operations. If a TFTP error is received, this field will be set to true after Self::tftp_error is updated.

make_callbacks: bool

When false, callbacks will not be made. When true, make callbacks to the PXE Base Code Callback Protocol. This field is reset to false by BaseCode::start if the PXE Base Code Callback Protocol is not available. It is reset to true by BaseCode::start if the PXE Base Code Callback Protocol is available.

ttl: u8

The “time to live” field of the IP header. This field is initialized to 16 by BaseCode::start and can be modified by BaseCode::set_parameters.

tos: u8

The type of service field of the IP header. This field is initialized to 0 by BaseCode::start, and can be modified with BaseCode::set_parameters.

station_ip: IpAddress

The device’s current IP address. This field is initialized to a zero address by Start(). This field is set when BaseCode::dhcp completes successfully. This field can also be set by BaseCode::set_station_ip. This field must be set to a valid IP address by either BaseCode::dhcp or BaseCode::set_station_ip before BaseCode::discover, BaseCode::udp_read, BaseCode::udp_write, BaseCode::arp and any of the TFTP/MTFTP operations are called.

subnet_mask: IpAddress

The device’s current subnet mask. This field is initialized to a zero address by BaseCode::start. This field is set when BaseCode::dhcp completes successfully. This field can also be set by BaseCode::set_station_ip. This field must be set to a valid subnet mask by either BaseCode::dhcp or BaseCode::set_station_ip before BaseCode::discover, BaseCode::udp_read, BaseCode::udp_write, BaseCode::arp or any of the TFTP/MTFTP operations are called.

dhcp_discover: Packet

Cached DHCP Discover packet. This field is zero-filled by the BaseCode::start function, and is set when BaseCode::dhcp completes successfully. The contents of this field can replaced by BaseCode::set_packets.

dhcp_ack: Packet

Cached DHCP Ack packet. This field is zero-filled by BaseCode::start, and is set when BaseCode::dhcp completes successfully. The contents of this field can be replaced by BaseCode::set_packets.

proxy_offer: Packet

Cached Proxy Offer packet. This field is zero-filled by BaseCode::start, and is set when BaseCode::dhcp completes successfully. The contents of this field can be replaced by BaseCode::set_packets.

pxe_discover: Packet

Cached PXE Discover packet. This field is zero-filled by BaseCode::start, and is set when BaseCode::discover completes successfully. The contents of this field can be replaced by BaseCode::set_packets.

pxe_reply: Packet

Cached PXE Reply packet. This field is zero-filled by BaseCode::start, and is set when BaseCode::discover completes successfully. The contents of this field can be replaced by the BaseCode::set_packets function.

pxe_bis_reply: Packet

Cached PXE BIS Reply packet. This field is zero-filled by BaseCode::start, and is set when BaseCode::discover completes successfully. This field can be replaced by BaseCode::set_packets.

ip_filter: IpFilter

The current IP receive filter settings. The receive filter is disabled and the number of IP receive filters is set to zero by BaseCode::start, and is set by BaseCode::set_ip_filter.

arp_cache_entries: u32

The number of valid entries in the ARP cache. This field is reset to zero by BaseCode::start.

arp_cache: [ArpEntry; 8]

Array of cached ARP entries.

route_table_entries: u32

The number of valid entries in the current route table. This field is reset to zero by BaseCode::start.

route_table: [RouteEntry; 8]

Array of route table entries.

icmp_error: IcmpError

ICMP error packet. This field is updated when an ICMP error is received and is undefined until the first ICMP error is received. This field is zero-filled by BaseCode::start.

tftp_error: TftpError

TFTP error packet. This field is updated when a TFTP error is received and is undefined until the first TFTP error is received. This field is zero-filled by the BaseCode::start function.