uefi::proto::network

Module snp

Simple Network Protocol

Provides a packet level interface to a network adapter. Once the adapter is initialized, the protocol provides services that allows packets to be transmitted and received.

No interface function must be called until SimpleNetwork.start is successfully called first.

Structs

InterruptStatus

Flags returned by get_interrupt_status to indicate which interrupts have fired on the interface since the last call.

NetworkMode

The Simple Network Mode

NetworkState

The state of a network interface.

NetworkStats

Network Statistics

ReceiveFlags

Flags to pass to receive_filters to enable/disable reception of some kinds of packets.

SimpleNetwork

The Simple Network Protocol