Struct uefi::proto::pi::mp::MpServices

#[repr(C)]
pub struct MpServices { /* fields omitted */ }

Protocol that provides services needed for multi-processor management.

Implementations

impl MpServices

pub fn get_number_of_processors(&self) -> Result<ProcessorCount>

Retrieves the number of logical processors and the number of enabled logical processors in the system.

pub fn get_processor_info(
&self,
processor_number: usize
) -> Result<ProcessorInformation>

Gets detailed information on the requested processor at the instant this call is made.

pub fn startup_all_aps(
&self,
single_thread: bool,
procedure: Procedure,
procedure_argument: *mut c_void,
timeout: Option<Duration>
) -> Result

Executes provided function on all APs in blocking mode.

pub fn startup_this_ap(
&self,
processor_number: usize,
procedure: Procedure,
procedure_argument: *mut c_void,
timeout: Option<Duration>
) -> Result

Executes provided function on a specific AP in blocking mode.

pub fn switch_bsp(
&self,
processor_number: usize,
enable_old_bsp: bool
) -> Result

Switches the requested AP to be the BSP from that point onward.

pub fn enable_disable_ap(
&self,
processor_number: usize,
enable_ap: bool,
healthy: Option<bool>
) -> Result

Enables or disables an AP from this point onward.

The healthy argument can be used to specify the new health status of the AP.

pub fn who_am_i(&self) -> Result<usize>

Gets the handle number of the caller processor.