uefi::proto::string

Module unicode_collation

The Unicode Collation Protocol.

This protocol is used in the boot services environment to perform lexical comparison functions on Unicode strings for given languages.

Structs

UnicodeCollation

The Unicode Collation Protocol.

Enums

StrConversionError

Errors returned by UnicodeCollation::str_lwr and UnicodeCollation::str_upr.