uefi::proto::tcg::v1

Struct PcrEvent

pub struct PcrEvent { /* private fields */ }

Entry in the EventLog.

Layout compatible with the C type TCG_PCR_EVENT.

Naming note: the spec refers to “event data” in two conflicting ways: the event_data field and the data hashed in the digest field. These two are independent; although the event data can be what is hashed in the digest field, it doesn’t have to be.

Implementations

impl PcrEvent

pub fn new_in_buffer<'buf>( buffer: &'buf mut [u8], pcr_index: PcrIndex, event_type: EventType, digest: Sha1Digest, event_data: &[u8], ) -> Result<&'buf mut Self, Option<usize>>

Create a new PcrEvent using a byte buffer for storage.

Errors

Returns Status::BUFFER_TOO_SMALL if the buffer is not large enough. The required size will be returned in the error data.

Returns Status::INVALID_PARAMETER if the event_data size is too large.

pub const fn pcr_index(&self) -> PcrIndex

PCR index for the event.

pub const fn event_type(&self) -> EventType

Type of event, indicating what type of data is stored in event_data.

pub const fn event_data(&self) -> &[u8]

Raw event data. The meaning of this data can be determined from the event_type.

Note that this data is independent of what is hashed digest.

pub const fn digest(&self) -> Sha1Digest

SHA-1 digest of the data hashed for this event.