uefi::proto::tcg::v2

Struct PcrEventInputs

pub struct PcrEventInputs { /* private fields */ }

Event type passed to Tcg::hash_log_extend_event.

Layout compatible with the C type EFI_TCG2_EVENT.

The TPM v1 spec uses a single generic event type for both creating a new event and reading an event from the log. The v2 spec splits this into two structs: EFI_TCG2_EVENT for creating events, and TCG_PCR_EVENT2 for reading events. To help clarify the usage, our API renames these types to PcrEventInputs and PcrEvent, respectively.

Implementations

impl PcrEventInputs

pub fn new_in_buffer<'buf>( buffer: &'buf mut [u8], pcr_index: PcrIndex, event_type: EventType, event_data: &[u8], ) -> Result<&'buf mut Self, Option<usize>>

Create a new PcrEventInputs using a byte buffer for storage.

Errors

Returns Status::BUFFER_TOO_SMALL if the buffer is not large enough. The required size will be returned in the error data.

Returns Status::INVALID_PARAMETER if the event_data size is too large.