uefi

Module runtime

UEFI runtime services.

These services are available both before and after exiting boot services. Note that various restrictions apply when calling runtime services functions after exiting boot services; see the “Calling Convention” section of the UEFI specification for details.

Structs

CapsuleBlockDescriptor

Descriptor that defines a scatter-gather list for passing a set of capsules to the firmware.

CapsuleFlags

Capsule update flags.

CapsuleHeader

Common header at the start of a capsule.

CapsuleInfo

Information about UEFI variable storage space returned by query_capsule_capabilities.

Daylight

A bitmask containing daylight savings time information.

ResetType

The type of system reset.

Time

Date and time representation.

TimeCapabilities

Real time clock capabilities.

TimeError

Error returned by Time methods. A bool value of true means the specified field is outside its valid range.

TimeParams

Input parameters for Time::new.

VariableAttributes

Flags describing the attributes of a variable.

VariableStorageInfo

Information about UEFI variable storage space returned by query_variable_info. Note that the data here is limited to a specific type of variable (as specified by the attributes argument to query_variable_info).

VariableVendor

Variable vendor GUID. This serves as a namespace for variables to avoid naming conflicts between vendors. The UEFI specification defines some special values, and vendors will define their own.

Enums

TimeByteConversionError

Error returned from failing to convert a byte slice into a Time.

Functions

delete_variable

Deletes a UEFI variable.

get_next_variable_key

Gets each variable key (name and vendor) one at a time.

get_time

Query the current time and date information.

get_time_and_caps

Query the current time and date information and the RTC capabilities.

get_variable

Gets the contents and attributes of a variable. The size of buf must be at least as big as the variable’s size, although it can be larger.

query_capsule_capabilities

Tests whether a capsule or capsules can be updated via update_capsule.

query_variable_info

Get information about UEFI variable storage space for the type of variable specified in attributes.

reset

Resets the computer.

set_time⚠

Sets the current local time and date information

set_variable

Sets the value of a variable. This can be used to create a new variable, update an existing variable, or (when the size of data is zero) delete a variable.

set_virtual_address_map⚠

Changes the runtime addressing mode of EFI firmware from physical to virtual. It is up to the caller to translate the old system table address to a new virtual address and provide it for this function.

update_capsule

Passes capsules to the firmware.

variable_exists

Checks if a variable exists.