uefi::runtime

Function get_next_variable_key

pub fn get_next_variable_key(
    name: &mut [u16],
    vendor: &mut VariableVendor,
) -> Result<(), Option<usize>>

Gets each variable key (name and vendor) one at a time.

This is used to iterate over variable keys. See [variable_keys] for a more convenient interface that requires the alloc feature.

To get the first variable key, name must be initialized to start with a null character. The vendor value is arbitrary. On success, the first variable’s name and vendor will be written out to name and vendor. Keep calling get_next_variable_key with the same name and vendor references to get the remaining variable keys.

All variable names should be valid strings, but this may not be enforced by firmware. To convert to a string, truncate at the first null and call CStr16::from_u16_with_nul.

Errors

Status::NOT_FOUND: indicates end of iteration, the last variable keys was retrieved by the previous call to get_next_variable_key.

Status::BUFFER_TOO_SMALL: name is not large enough. The required size (in u16 characters, not bytes) will be returned in the error data.

Status::INVALID_PARAMETER: name does not contain a null character, or the name and vendor are not an existing variable.

Status::DEVICE_ERROR: variable could not be read due to a hardware error.

Status::UNSUPPORTED: this platform does not support variable storage after exiting boot services.