uefi::runtime

Function query_capsule_capabilities

pub fn query_capsule_capabilities(
    capsule_header_array: &[&CapsuleHeader],
) -> Result<CapsuleInfo>

Tests whether a capsule or capsules can be updated via update_capsule.

See CapsuleInfo for details of the information returned.

Errors

Status::OUT_OF_RESOURCES: before exiting boot services, indicates the capsule is compatible with the platform but there are insufficient resources to complete the update. After exiting boot services, indicates the capsule is compatible with the platform but can only be processed before exiting boot services.

Status::UNSUPPORTED: either the capsule type is not supported by this platform, or the platform does not support capsule updates after exiting boot services.