uefi::runtime

Function query_variable_info

pub fn query_variable_info(
    attributes: VariableAttributes,
) -> Result<VariableStorageInfo>

Get information about UEFI variable storage space for the type of variable specified in attributes.

This operation is only supported starting with UEFI 2.0.

See VariableStorageInfo for details of the information returned.

Errors

Status::INVALID_PARAMETER: invalid combination of variable attributes.

Status::UNSUPPORTED: the combination of variable attributes is not supported on this platform, or the UEFI version is less than 2.0.