uefi::runtime

Function set_variable

pub fn set_variable(
    name: &CStr16,
    vendor: &VariableVendor,
    attributes: VariableAttributes,
    data: &[u8],
) -> Result

Sets the value of a variable. This can be used to create a new variable, update an existing variable, or (when the size of data is zero) delete a variable.

Warnings

The Status::WARN_RESET_REQUIRED warning will be returned when using this function to transition the Secure Boot mode to setup mode or audit mode if the firmware requires a reboot for that operation.

Errors

Status::INVALID_PARAMETER: invalid attributes, name, or vendor.

Status::OUT_OF_RESOURCES: not enough storage is available to hold the variable.

Status::WRITE_PROTECTED: variable is read-only.

Status::SECURITY_VIOLATION: variable could not be written due to an authentication error.

Status::NOT_FOUND: attempted to update a non-existent variable.

Status::UNSUPPORTED: this platform does not support variable storage after exiting boot services.