uefi::runtime

Struct Time

pub struct Time(/* private fields */);

Date and time representation.

Implementations

impl Time

pub fn new(params: TimeParams) -> Result<Self, TimeError>

Create a Time value. If a field is not in the valid range, TimeError is returned.

pub const fn invalid() -> Self

Create an invalid Time with all fields set to zero. This can be used with FileInfo to indicate a field should not be updated when calling File::set_info.

pub fn is_valid(&self) -> Result<(), TimeError>

Ok() if all fields are within valid ranges, Err(TimeError) otherwise.

pub const fn year(&self) -> u16

Query the year.

pub const fn month(&self) -> u8

Query the month.

pub const fn day(&self) -> u8

Query the day.

pub const fn hour(&self) -> u8

Query the hour.

pub const fn minute(&self) -> u8

Query the minute.

pub const fn second(&self) -> u8

Query the second.

pub const fn nanosecond(&self) -> u32

Query the nanosecond.

pub const fn time_zone(&self) -> Option<i16>

Query the time offset in minutes from UTC, or None if using local time.

pub const fn daylight(&self) -> Daylight

Query the daylight savings time information.