Struct uefi::Completion

#[must_use]
pub struct Completion<T> { /* fields omitted */ }

This type is used when an UEFI operation has completed, but some non-fatal problems (UEFI warnings) may have been encountered along the way

Implementations

impl<T> Completion<T>

pub fn new(status: Status, result: T) -> Self

Build a completion from a non-error status and a function result

pub fn status(&self) -> Status

Extract the status of this completion

pub fn split(self) -> (Status, T)

Split this completion into its inner status and result data

pub fn log(self) -> T

Access the inner value, logging the warning if there is any

pub fn unwrap(self) -> T

Assume that no warning occured, panic if not

pub fn expect(self, msg: &str) -> T

Assume that no warning occured, panic with provided message if not

pub fn map<U>(self, f: impl FnOnce(T) -> U) -> Completion<U>

Transform the inner value without unwrapping the Completion

pub fn with_status(self, extra_status: Status) -> Self

Merge this completion with a success or warning status

Since this type only has storage for one warning, if two warnings must be stored, one of them will be spilled into the logs.