uefi

Module system

Functions for accessing fields of the system table.

Some of these functions use a callback argument rather than returning a reference to the field directly. This pattern is used because some fields are allowed to change, and so a static lifetime cannot be used.

Some functions can only be called while boot services are active, and will panic otherwise. See each function’s documentation for details.

Functions

firmware_revision

Get the firmware revision.

firmware_vendor

Get the firmware vendor string.

uefi_revision

Get the revision of the system table, which is defined to be the revision of the UEFI specification implemented by the firmware.

with_config_table

Call f with a slice of ConfigTableEntry. Each entry provides access to a vendor-specific table.

with_stderr

Call f with the Output protocol attached to stderr.

with_stdin

Call f with the Input protocol attached to stdin.

with_stdout

Call f with the Output protocol attached to stdout.