Module uefi::table

Standard UEFI tables.

Modules

boot

UEFI services available during boot.

cfg

Configuration table utilities.

runtime

UEFI services available at runtime, even after the OS boots.

Structs

Boot

Marker struct associated with the boot view of the UEFI System Table

Header

All standard UEFI tables begin with a common header.

Revision

A revision of the UEFI specification.

Runtime

Marker struct associated with the run-time view of the UEFI System Table

SystemTable

UEFI System Table interface

Traits

Table

Common trait implemented by all standard UEFI tables.