Struct uefi::table::runtime::RuntimeServices

#[repr(C)]
pub struct RuntimeServices { /* fields omitted */ }

Contains pointers to all of the runtime services.

This table, and the function pointers it contains are valid even after the UEFI OS loader and OS have taken control of the platform.

Implementations

impl RuntimeServices

pub fn get_time(&self) -> Result<Time>

Query the current time and date information

pub fn get_time_and_caps(&self) -> Result<(Time, TimeCapabilities)>

Query the current time and date information and the RTC capabilities

pub unsafe fn set_time(&mut self, time: &Time) -> Result

Sets the current local time and date information

During runtime, if a PC-AT CMOS device is present in the platform, the caller must synchronize access to the device before calling set_time.

Safety

Undefined behavior could happen if multiple tasks try to use this function at the same time without synchronisation.

pub unsafe fn set_virtual_address_map(
&self,
map: &mut [MemoryDescriptor]
) -> Result

Changes the runtime addressing mode of EFI firmware from physical to virtual.

Safety

Setting new virtual memory map is unsafe and may cause undefined behaviors.

pub fn reset(&self, rt: ResetType, status: Status, data: Option<&[u8]>) -> !

Resets the computer.