uefi

Trait StatusExt

pub trait StatusExt {
    // Required methods
    fn to_result(self) -> Result;
    fn to_result_with_val<T>(self, val: impl FnOnce() -> T) -> Result<T, ()>;
    fn to_result_with_err<ErrData: Debug>(
        self,
        err: impl FnOnce(Status) -> ErrData,
    ) -> Result<(), ErrData>;
    fn to_result_with<T, ErrData: Debug>(
        self,
        val: impl FnOnce() -> T,
        err: impl FnOnce(Status) -> ErrData,
    ) -> Result<T, ErrData>;
}

Extension trait which provides some convenience methods for Status.

Required Methods

fn to_result(self) -> Result

Converts this status code into a uefi::Result.

If the status does not indicate success, the status representing the specific error code is embedded into the Err variant of type uefi::Error.

fn to_result_with_val<T>(self, val: impl FnOnce() -> T) -> Result<T, ()>

Converts this status code into a uefi::Result with a given Ok value.

If the status does not indicate success, the status representing the specific error code is embedded into the Err variant of type uefi::Error.

fn to_result_with_err<ErrData: Debug>( self, err: impl FnOnce(Status) -> ErrData, ) -> Result<(), ErrData>

Converts this status code into a uefi::Result with a given Err payload.

If the status does not indicate success, the status representing the specific error code is embedded into the Err variant of type uefi::Error.

fn to_result_with<T, ErrData: Debug>( self, val: impl FnOnce() -> T, err: impl FnOnce(Status) -> ErrData, ) -> Result<T, ErrData>

Convert this status code into a result with a given Ok value and Err payload.

If the status does not indicate success, the status representing the specific error code is embedded into the Err variant of type uefi::Error.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".