uefi

Type Alias Result

pub type Result<Output = (), ErrData = ()> = Result<Output, Error<ErrData>>;

Return type of most UEFI functions. Both success and error payloads are optional.

Almost all UEFI operations provide a status code as an output which indicates either success, a warning, or an error. This type alias maps Status::SUCCESS to the Ok variant (with optional Output data), and maps both warning and error statuses to the Err variant of type Error, which may carry optional inner ErrData.

Warnings are treated as errors by default because they generally indicate an abnormal situation.

Some convenience methods are provided by the ResultExt trait.

Aliased Type

pub enum Result<Output = (), ErrData = ()> {
    Ok(Output),
    Err(Error<ErrData>),
}

Variants

Ok(Output)

Contains the success value

Err(Error<ErrData>)

Contains the error value