uefi::boot

Function open_protocol_exclusive

pub fn open_protocol_exclusive<P: ProtocolPointer + ?Sized>(
    handle: Handle,
) -> Result<ScopedProtocol<P>>

Opens a protocol interface for a handle in exclusive mode.

If successful, a ScopedProtocol is returned that will automatically close the protocol interface when dropped.

Errors

Status::UNSUPPORTED: the handle does not support the protocol.

Status::ACCESS_DENIED: the protocol is already open in a way that is incompatible with the new request.