Module uefi::data_types

Data type definitions

This module defines the basic data types that are used throughout uefi-rs

Re-exports

pub use self::chars::Char16;

pub use self::chars::Char8;

Modules

chars
UEFI character handling

Structs

CStr8
A Latin-1 null-terminated string

CStr16
An UCS-2 null-terminated string

Event
Handle to an event structure

Guid
A globally unique identifier

Handle
Opaque handle to an UEFI entity (protocol, image…)

Traits

Align
Trait for querying the alignment of a struct

Identify
Several entities in the UEFI specification can be referred to by their GUID, this trait is a building block to interface them in uefi-rs.

Attribute Macros

unsafe_guid
unsafe_guid attribute macro, implements the Identify trait for any type (mostly works like a custom derive, but also supports type aliases)