Module uefi::proto::media::file

This module provides the FileHandle structure as well as the more specific RegularFile and Directory structures. This module also provides the File trait for opening, querying, creating, reading, and writing files.

Usually a file system implementation will return a “root” directory, representing / on that volume. With that directory, it is possible to enumerate and open all the other files on that volume.

Structs

Directory
A FileHandle that is also a directory.

FileAttribute
Attributes describing the properties of a file on the file system.

FileHandle
An opaque handle to some contiguous block of data on a volume.

NamedFileProtocolInfo
Dynamically sized FileProtocolInfo with a header and an UCS-2 name

RegularFile
A FileHandle that is also a regular (data) file.

Enums

FileMode
Usage flags describing what is possible to do with the file.

FileType
Disambiguates the file type. Returned by File::into_type().

Traits

File
Common interface to FileHandle, RegularFile, and Directory.

FileProtocolInfo
Common trait for data structures that can be used with File::set_info() or File::get_info().

FromUefi
Trait for going from an UEFI-originated pointer to a Rust reference

Type Definitions

FileInfo
Generic file information

FileSystemInfo
System volume information

FileSystemVolumeLabel
System volume label