Module uefi::table::cfg

Configuration table utilities.

The configuration table is an array of GUIDs and pointers to extra system tables.

For example, it can be used to find the ACPI tables.

This module contains the actual entries of the configuration table, as well as GUIDs for many known vendor tables.

Structs

ConfigTableEntry
Contains a set of GUID / pointer for a vendor-specific table.

MemoryProtectionAttribute
Flags describing memory protection.

PropertiesTable
This table contains additional information about the UEFI implementation.

Constants

ACPI2_GUID
Entry pointing to the ACPI 2 RSDP.

ACPI_GUID
Entry pointing to the old ACPI 1 RSDP.

DEBUG_IMAGE_INFO_GUID
Pointer to the debug image info table.

DXE_SERVICES_GUID
Table which provides Driver eXecution Environment services.

HAND_OFF_BLOCK_LIST_GUID
Hand-off Blocks are used to pass data from the early pre-UEFI environment to the UEFI drivers.

LZMA_COMPRESS_GUID
LZMA-compressed filesystem.

MEMORY_STATUS_CODE_RECORD_GUID
Used to identify Hand-off Blocks which store status codes reported during the pre-UEFI environment.

MEMORY_TYPE_INFORMATION_GUID
Table used in the early boot environment to record memory ranges.

PROPERTIES_TABLE_GUID
GUID of the UEFI properties table.

SMBIOS3_GUID
Entry pointing to the SMBIOS 3.0 table.

SMBIOS_GUID
Entry pointing to the SMBIOS 1.0 table.

TIANO_COMPRESS_GUID
A custom compressed filesystem used by the Tiano UEFI implementation.